requirements.txt
aiofiles==0.5.0
aiohttp==3.9.5
aiohttp-jinja2==1.5
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.13.3
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.5.7
bcrypt==3.2.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.4
cryptography==45.0.5
disposable-email-domains==0.0.133
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastapi-mail==1.5.0
fastapi-sso==0.17.0
firebase-admin==6.9.0
frozenlist==1.3.3
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==1.1.0
Jinja2==3.1.6
Mako==1.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.0.2
oauthlib==3.2.2
orjson==3.7.11
passlib==1.7.4
promise==2.3
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.7.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.1.1
python-editor==1.0.4
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.5.5
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rsa==4.9.1
Rx==1.6.1
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
toml==0.10.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
watchgod==0.7
websockets==15.0.1
yarl==1.8.1
debugpy
pytest
pytest-asyncio